# pkg/serve/testdata/openapi.yaml
...
openapi: 3.1.0
info:
  title: looprig serve — HTTP session API
  version: "1"
  description: >
    HTTP surface over a live/durable agent session: lifecycle (create/restore),
    control (input/interrupt/gate response), reads (list/status/journal), and a
    Server-Sent-Events live stream. The error body is a nested envelope
    {"error":{code,message,retryable}}. All routes are under /v1.
servers:
  - url: /v1
...
paths:
...
  /sessions/{sid}/restore:
    post:
      tags: [lifecycle]
      summary: Attach to a live session, or rebuild a prior one from durable history.
      operationId: restoreSession
      description: >
        Idempotent. If {sid} is already live in this process the rig is NOT touched
        and the response carries restored=false; the caller keeps every existing
        event subscription. Otherwise the session is rebuilt from durable history,
        attached, and restored=true is returned. Rebuilding an already-live session
        would either fail on its exclusive workspace lease or mint a second runtime
        over the same journal and orphan existing subscribers, so the live case
        short-circuits before the rig is consulted.
      parameters:
        - $ref: "#/components/parameters/Sid"
      responses:
        "200":
          description: Session attached (already live) or rebuilt and reattached.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RestoreResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  parameters:
    Sid:
      name: sid
      in: path
      required: true
      description: Canonical 8-4-4-4-12 UUID session id.
      schema: { $ref: "#/components/schemas/UUID" }
...
  responses:
    BadRequest:
      description: Malformed path/query parameter or request body.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Unauthorized:
      description: Authentication required or failed (when auth is configured).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: No such session (or gate).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Internal:
      description: An unexpected server-side failure (cause is logged, never in the body).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
    UUID:
      type: string
      pattern: "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$"
...
    RestoreResponse:
      type: object
      additionalProperties: false
      required: [session_id, restored]
      properties:
        session_id: { $ref: "#/components/schemas/UUID" }
        restored:
          type: boolean
          description: >
            true when the rig rebuilt the session from durable history; false when
            an already-live session was reused (attach). Both are 200.
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, retryable]
          properties:
            code:
              type: string
              enum:
                - internal
                - unauthorized
                - invalid_body
                - invalid_parameter
                - session_not_found
                - idempotency_conflict
                - gate_not_found
                - gate_action_invalid
                - gate_kind_mismatch
                - gate_not_ready
                - gate_capacity
            message: { type: string }
            retryable: { type: boolean }